framework/task_manager: refactoring task manager
seperate source codes to each file per a type of request
