+++ apps/builtin/builtin_list.c
@@ -103,5 +103,5 @@ void register_examples_cmds(void)
 
 int get_builtin_list_cnt(void)
 {
-	return sizeof(builtin_list) / sizeof(builtin_list[0]);
+	return sizeof(builtin_list) / sizeof(builtin_list[0]) - 1;
 }
