+++ framework/src/task_manager/Make.defs
@@ -17,7 +17,9 @@
 ###########################################################################
 
 ifeq ($(CONFIG_TASK_MANAGER),y)
-CSRCS += task_manager.c taskmgr_interface.c
+CSRCS += task_manager_register.c task_manager_unregister.c task_manager_start.c task_manager_terminate.c task_manager_pause.c task_manager_resume.c
+CSRCS += task_manager_getinfo.c task_manager_unicast.c task_manager_cleaninfo.c task_manager_sethandler.c task_manager_state.c task_manager_permission.c
+CSRCS += task_manager_core.c task_manager_interface.c
 
 DEPPATH += --dep-path src/task_manager
 VPATH += :src/task_manager
