+++ framework/src/task_manager/task_manager_internal.h
@@ -19,6 +19,9 @@
 #ifndef __TASK_MANAGER_INTERNAL_H__
 #define __TASK_MANAGER_INTERNAL_H__
 
+#include <stdbool.h>
+#include <unistd.h>
+
 /* Command Types */
 #define TASKMGT_REGISTER_TASK        0
 #define TASKMGT_REGISTER_THREAD      1
@@ -38,9 +41,60 @@
 /* Message Queue Values */
 #define TM_MQ_PRIO   50
 #define TM_PUBLIC_MQ "tm_public_mq"
+#define TM_PRIVATE_MQ "tm_priv_mq"
 
 /* Wrapper of allocation APIs */
 #define TM_ALLOC(a)  malloc(a)
 #define TM_FREE(a)   free(a)
+
+struct task_list_s {
+	int pid;
+	void *addr;
+};
+typedef struct task_list_s task_list_t;
+
+struct task_list_data_s {
+	int builtin_idx;
+	int pid;
+	int tm_gid;
+	int status;
+	int permission;
+};
+typedef struct task_list_data_s task_list_data_t;
+
+struct tm_request_s {
+	int cmd;
+	int caller_pid;
+	int handle;
+	int timeout;
+	char *q_name;
+	void* data;
+};
+typedef struct tm_request_s tm_request_t;
+
+struct tm_response_s {
+	int status;
+	task_info_list_t *data;
+};
+typedef struct tm_response_s tm_response_t;
+
+#define IS_INVALID_HANDLE(i) (i < 0 || i >= CONFIG_TASK_MANAGER_MAX_TASKS)
+
+#define TASK_LIST_ADDR(handle)   ((task_list_data_t *)tm_task_list[handle].addr)
+#define TASK_PID(handle)         tm_task_list[handle].pid
+#define TASK_BUILTIN_IDX(handle) TASK_LIST_ADDR(handle)->builtin_idx
+#define TASK_TM_GID(handle)      TASK_LIST_ADDR(handle)->tm_gid
+#define TASK_STATUS(handle)      TASK_LIST_ADDR(handle)->status
+#define TASK_PERMISSION(handle)  TASK_LIST_ADDR(handle)->permission
+
+extern task_list_t tm_task_list[CONFIG_TASK_MANAGER_MAX_TASKS];
+
+int taskmgr_send_request(tm_request_t *request_msg);
+int taskmgr_send_response(char *q_name, tm_response_t *response_msg);
+int taskmgr_receive_response(char *q_name, tm_response_t *response_msg, int timeout);
+
+bool taskmgr_is_permitted(int handle, pid_t pid);
+int taskmgr_get_task_state(int handle);
+int taskmgr_get_drvfd(void);
 
 #endif
