+++ os/include/tinyara/fs/ioctl.h
@@ -358,20 +358,13 @@
 #define _TMIOCVALID(c)   (_IOC_TYPE(c) == _TMBASE)
 #define _TMIOC(nr)       _IOC(_TMBASE, nr)
 
-#define TMIOC_REGISTER_TASK        _TMIOC(0x0001)
-#define TMIOC_REGISTER_THREAD      _TMIOC(0x0002)
-#define TMIOC_START                _TMIOC(0x0003)
-#define TMIOC_TERMINATE            _TMIOC(0x0004)
-#define TMIOC_RESTART              _TMIOC(0x0005)
-#define TMIOC_PAUSE                _TMIOC(0x0006)
-#define TMIOC_RESUME               _TMIOC(0x0007)
-#define TMIOC_SCAN                 _TMIOC(0x0008)
-#define TMIOC_UNICAST              _TMIOC(0x0009)
-#define TMIOC_BROADCAST            _TMIOC(0x000A)
-#define TMIOC_SCAN_NAME            _TMIOC(0x000B)
-#define TMIOC_SCAN_HANDLE          _TMIOC(0x000C)
-#define TMIOC_SCAN_GROUP           _TMIOC(0x000D)
-#define TMIOC_UNREGISTER_TASK      _TMIOC(0x000E)
+#define TMIOC_START                _TMIOC(0x0001)
+#define TMIOC_PAUSE                _TMIOC(0x0002)
+#define TMIOC_RESUME               _TMIOC(0x0003)
+#define TMIOC_RESTART              _TMIOC(0x0004)
+#define TMIOC_UNICAST              _TMIOC(0x0005)
+#define TMIOC_BROADCAST            _TMIOC(0x0006)
+#define TMIOC_CHECK_ALIVE          _TMIOC(0x0007)
 
 /****************************************************************************
  * Public Type Definitions
